La till Vm för auction details, ändrade bid size till double, la till bidVm. Bids visas inte ännu i auction details.

+++ ProjectApp/Core/Bid.cs
@@ -4,16 +4,16 @@ public class Bid
 {
     public int id { get; set; }
     public string username { get; set; }
-    public int bidSize { get; set; }
+    public double bidSize { get; set; }
     public DateTime bidTime { get; set; }
 
-    public Bid(string username, int bidSize, DateTime bidTime)
+    public Bid(string username, double bidSize, DateTime bidTime)
     {
         this.username = username;
         this.bidSize = bidSize;
         this.bidTime = bidTime;
     }
-    public Bid(int id, string username, int bidSize, DateTime bidTime)
+    public Bid(int id, string username, double bidSize, DateTime bidTime)
     {
         this.id = id;
         this.username = username;
